+++ app/recipes/apis/recipe_name.py
@@ -7,11 +7,11 @@ from ..models import RecipeName
 
 
 class RecipeNameListCreateView(generics.ListCreateAPIView):
-    # queryset = RecipeName.objects.all()
+    queryset = RecipeName.objects.all()
     serializer_class = RecipeNameSerializer
 
     # Data caching by Redis
-    queryset = cache.get_or_set('recipenames', RecipeName.objects.all().values(), 3600)
+    # queryset = cache.get_or_set('recipenames', RecipeName.objects.all().values(), 3600)
 
     permission_classes = (
         permissions.IsAuthenticatedOrReadOnly,
