+++ app/recipes/models/__init__.py
@@ -1,4 +1,5 @@
-from .intermediate import *
+from .bookmarkedrecipe import *
+from .likedrecipe import *
 from .manytomany import *
 from .manytoone import *
 from .onetoone import *
